+++ CH05/README.md
@@ -86,14 +86,26 @@ $$
 > 输入：训练数据集$D$, 特征集$A$，阈值$\epsilon$
 > 输出：决策树$T$
 >
-> 1. $D$属于同一类C_k -> T为单节点树，类Ck作为该节点的类标记，返回T
-> 1. A是空集->T为单节点树，实例数最多的作为该节点类标记，返回T
-> 1. 计算g, 选择信息增益最大的特征Ag
-> 1. 如果Ag的信息增益小于ε，T为单节点树，D中实例数最大的类Ck作为类标记，返回T
-> 1. Ag划分若干非空子集Di，
-> 1. Di训练集，A-{Ag}为特征集，递归调用前面步骤，得到Ti，返回Ti
+> 1. $D$属于同一类$C_k$，$ T$为单节点树，类$C_k$作为该节点的类标记，返回$T$
+> 1. $A$是空集->T为单节点树，实例数最多的作为该节点类标记，返回T
+> 1. 计算$g$, 选择信息增益最大的特征$A_g$
+> 1. 如果$A_g$的信息增益小于$\epsilon$，$T$为单节点树，$D$中实例数最大的类$C_k$作为类标记，返回$T$
+> 1. $A_g$划分若干非空子集$D_i$，
+> 1. $D_i$训练集，$A-A_g$为特征集，递归调用前面步骤，得到$T_i$，返回$T_i$
 
-### 算法5.3
+
+
+### 算法5.3 C4.5生成算法
+
+### 算法5.4 树的剪枝算法
+
+### 算法5.5 最小二乘回归树生成算法
+
+### 算法5.6 CART生成算法
+
+### 算法5.7 CART剪枝算法
+
+
 
 ## 例子
