+++ CH05/unit_test.py
@@ -28,7 +28,7 @@ class TestDT(unittest.TestCase):
         logger.info(hd)
 
     def test_e53(self):
-        
+        pass
 
     def test_e54(self):
         pass
